Create new `useDropdownItem` composable
This composable defines common methods that can be shared across different dropdown menu item types.
